refactor(kubernautagent): split session/manager.go into focused files
Extracts StartInteractiveSession/LaunchDeferredInvestigation/
UpgradeToInteractive/CancelInvestigation/SuspendInvestigation/
TransitionToUserDriving/ForceTransitionToUserDriving into
manager_interactive.go, FindByRemediationID/Subscribe/GetSession/
CompleteUserDriving/ForceCompleteByRemediationID and other read/lookup
methods into manager_query.go, and storePartialResult/recoverPanic/
emitCompleteEvent/EmitSessionEndedByRR/EmitAccessDenied/
recordSessionMetrics/emitSessionEvent into manager_events.go. Reduces
manager.go from 1085 to 309 lines (NewManager + autonomous launch flow
only), no behavioral changes.
